refactor: split session interaction handlers
